tools/validate_config/requirements.txt
pyyaml>=6.0.1
jsonschema>=4.20.0 